JAMES-1607 extract spring specific stuff out of  MailboxCopierManagement
