+++ container/mailbox-adapter/src/main/java/org/apache/james/adapter/mailbox/MailboxManagerResolverException.java
@@ -16,8 +16,12 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
+package org.apache.james.adapter.mailbox;
 
-package org.apache.james.utils;
+public class MailboxManagerResolverException extends RuntimeException {
 
-public interface MailboxCopierManagementMBean extends org.apache.james.container.spring.mailbox.MailboxCopierManagementMBean {
+    public MailboxManagerResolverException(Throwable cause) {
+        super(cause);
+    }
+
 }
